requirements.txt
django
pandas
numpy
scikit-learn
scipy
nltk
gensim
pymongo
django-environ
six
django-utils-six
crispy-bootstrap5
django-crispy-forms
psycopg2-binary
channels
openai
channels_redis
uwsgi
celery==5.2.7
redis
